+++ packages/bot-utils/src/utils/array/index.ts
@@ -1,2 +1,4 @@
 export * from './sort';
 export * from './unique';
+export * from './cut';
+export * from './remove-nan-values';