+++ packages/openai-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @kkkaoru/openai-utils
 
+## 1.1.3
+
+### Patch Changes
+
+- remove utils and use bot utils
+
 ## 1.1.2
 
 ### Patch Changes